backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"
youtube-transcript-api = "^0.6.2"
google-api-python-client = "^2.125.0"
sqlalchemy = "^2.0.29"
aiomysql = "^0.2.0"
pytest-asyncio = "^0.23.6"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
